Create OSI Java packaging
