+++ src/main/java/com/vaticle/typedb/osi/config/ConfigurationHandler.java
@@ -14,18 +14,17 @@
  * limitations under the License.
  */
 
-package com.vaticle.typedb.osi.io;
+package com.vaticle.typedb.osi.config;
 
-import org.junit.Test;
+public enum ConfigurationHandler {
+    ATTRIBUTES,
+    ENTITIES,
+    RELATIONS,
+    APPEND_ATTRIBUTE,
+    APPEND_ATTRIBUTE_OR_INSERT_THING;
 
-public class ErrorLoggerTest {
-
-    @Test
-    public void errorFileLogger() {
-        FileLogger logger = FileLogger.getLogger();
-        logger.logMalformed("entities.tsv", "shucks! There was a malformed row error");
-        logger.logUnavailable("entities.tsv", "shucks! There was a connection error");
-        logger.logInvalid("entities.tsv", "shucks! There was a invalid row error");
+    @Override
+    public String toString() {
+        return super.toString().toLowerCase();
     }
-
 }
