+++ src/main/java/com/vaticle/typedb/osi/generator/Generator.java
@@ -14,18 +14,11 @@
  * limitations under the License.
  */
 
-package com.vaticle.typedb.osi.io;
+package com.vaticle.typedb.osi.generator;
 
-import org.junit.Test;
+import com.vaticle.typedb.client.api.TypeDBTransaction;
 
-public class ErrorLoggerTest {
-
-    @Test
-    public void errorFileLogger() {
-        FileLogger logger = FileLogger.getLogger();
-        logger.logMalformed("entities.tsv", "shucks! There was a malformed row error");
-        logger.logUnavailable("entities.tsv", "shucks! There was a connection error");
-        logger.logInvalid("entities.tsv", "shucks! There was a invalid row error");
-    }
-
+public interface Generator {
+    void write(TypeDBTransaction tx, String[] row);
+    char getFileSeparator();
 }
