+++ src/main/java/com/vaticle/typedb/osi/type/AttributeValueType.java
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package com.vaticle.typedb.osi.io;
+package com.vaticle.typedb.osi.type;
 
-import org.junit.Test;
+public enum AttributeValueType {
+    STRING,
+    LONG,
+    DOUBLE,
+    BOOLEAN,
+    DATETIME,
+    INVALID;
 
-public class ErrorLoggerTest {
-
-    @Test
-    public void errorFileLogger() {
-        FileLogger logger = FileLogger.getLogger();
-        logger.logMalformed("entities.tsv", "shucks! There was a malformed row error");
-        logger.logUnavailable("entities.tsv", "shucks! There was a connection error");
-        logger.logInvalid("entities.tsv", "shucks! There was a invalid row error");
-    }
-
+//    @Override
+//    public String toString() {
+//        return super.toString().toLowerCase();
+//    }
 }
