--- src/main/java/config/ConfigurationHandler.java
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2021 Bayer AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package config;
-
-public enum ConfigurationHandler {
-    ATTRIBUTES,
-    ENTITIES,
-    RELATIONS,
-    APPEND_ATTRIBUTE,
-    APPEND_ATTRIBUTE_OR_INSERT_THING;
-
-    @Override
-    public String toString() {
-        return super.toString().toLowerCase();
-    }
-}
