+++ src/main/resources/logback.xml
@@ -2,5 +2,5 @@
     <logger name="io.netty.util.internal" level="ERROR"/>
     <logger name="io.netty" level="OFF"/>
     <logger name="io.grpc" level="OFF"/>
-    <root level="OFF"/>
+    <root level="DEBUG"/>
 </configuration>
