+++ src/test/java/cli/TypeDBLoaderCLITest.java
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package cli;
+package com.vaticle.typedb.osi.cli;
 
 import com.vaticle.typedb.client.api.TypeDBClient;
 import com.vaticle.typedb.client.api.TypeDBSession;
 import com.vaticle.typedb.client.api.TypeDBTransaction;
 import com.vaticle.typeql.lang.TypeQL;
-import loader.TypeDBLoader;
+import com.vaticle.typedb.osi.loader.TypeDBLoader;
 import org.junit.Test;
-import util.TypeDBUtil;
+import com.vaticle.typedb.osi.util.TypeDBUtil;
 
 import static org.junit.Assert.*;
