+++ src/test/java/config/ConfigurationValidationTest.java
@@ -15,10 +15,10 @@
  */
 
 
-package config;
+package com.vaticle.typedb.osi.config;
 
-import util.TypeDBUtil;
-import util.Util;
+import com.vaticle.typedb.osi.util.TypeDBUtil;
+import com.vaticle.typedb.osi.util.Util;
 import com.vaticle.typedb.client.api.TypeDBClient;
 import com.vaticle.typedb.client.api.TypeDBSession;
 import org.apache.logging.log4j.LogManager;
