+++ src/test/java/generator/AppendAttributeOrInsertThingGeneratorTest.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package generator;
+package com.vaticle.typedb.osi.generator;
 
 
-import config.Configuration;
-import util.TypeDBUtil;
-import util.Util;
+import com.vaticle.typedb.osi.config.Configuration;
+import com.vaticle.typedb.osi.util.TypeDBUtil;
+import com.vaticle.typedb.osi.util.Util;
 import com.vaticle.typedb.client.api.TypeDBClient;
 import com.vaticle.typedb.client.api.TypeDBSession;
 import com.vaticle.typeql.lang.query.TypeQLInsert;
